interface/dialog: fix iteration variable type compared to vlc_array_count
Following the changes introduced by 7ecd927, the affected lines would
result in a warning due to signed vs unsigned integer comparision.

+++ src/interface/dialog.c
@@ -169,7 +169,7 @@ libvlc_InternalDialogInit(libvlc_int_t *p_libvlc)
 static int
 dialog_get_idx_locked(vlc_dialog_provider *p_provider, vlc_dialog_id *p_id)
 {
-    for (int i = 0; i < vlc_array_count(&p_provider->dialog_array); ++i)
+    for (size_t i = 0; i < vlc_array_count(&p_provider->dialog_array); ++i)
     {
         if (p_id == vlc_array_item_at_index(&p_provider->dialog_array, i))
             return i;
@@ -231,7 +231,7 @@ dialog_remove_locked(vlc_dialog_provider *p_provider, vlc_dialog_id *p_id)
 static void
 dialog_clear_all_locked(vlc_dialog_provider *p_provider)
 {
-    for (int i = 0; i < vlc_array_count(&p_provider->dialog_array); ++i)
+    for (size_t i = 0; i < vlc_array_count(&p_provider->dialog_array); ++i)
     {
         vlc_dialog_id *p_id =
             vlc_array_item_at_index(&p_provider->dialog_array, i);
